+++ packages/plugin-packages/mashroom-portal-remote-app-registry/src/js/registry/mashroom-bootstrap-portal-app-registry.ts
@@ -5,11 +5,11 @@ import context from '../context';
 import healthProbe from '../health/health-probe';
 import {registerRemoteAppMetrics, unregisterRemoteAppMetrics} from '../metrics/remote-app-metrics';
 
-import type {MashroomRemotePortalAppRegistryBootstrapFunction} from '@mashroom/mashroom-portal/type-definitions';
+import type {MashroomPortalAppRegistryBootstrapFunction} from '@mashroom/mashroom-portal/type-definitions';
 import type {MashroomLoggerFactory} from '@mashroom/mashroom/type-definitions';
 import type {MashroomPortalRemoteAppEndpointService} from '../../../type-definitions';
 
-const bootstrap: MashroomRemotePortalAppRegistryBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
+const bootstrap: MashroomPortalAppRegistryBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
     const { remotePortalAppUrls } = pluginConfig;
     const {loggerFactory, services, serverConfig} = pluginContextHolder.getPluginContext();
 
@@ -43,7 +43,7 @@ async function registerRemotePortalAppsFromConfigFile(configFilePath: string, se
 
     if (fs.existsSync(configFilePath)) {
         logger.info(`Loading remote portal app URLs from: ${configFilePath}`);
-         
+
         const remotePortalAppModule = require(configFilePath);
         const remotePortalAppData = remotePortalAppModule.default ?? remotePortalAppModule;
         const remotePortalAppURLs: Array<string> = Array.isArray(remotePortalAppData) ? remotePortalAppData : (remotePortalAppData.remotePortalApps || []);