+++ packages/plugin-packages/mashroom-portal-sandbox-app/src/js/containers/PortalAppSelectionContainer.tsx
@@ -6,7 +6,7 @@ import type {State} from '../types';
 
 const mapStateToProps = (state: State) => {
     return {
-        availablePortalApps: state.availablePortalApps,
+        knownPortalApps: state.knownPortalApps,
         appLoadingError: state.appLoadingError
     };
 };