+++ packages/plugin-packages/mashroom-portal/src/backend/plugins/loader/mashroom-bootstrap-portal-app-registry-plugin-loader.ts
@@ -1,11 +1,11 @@
 
 import context from '../../context/global-portal-context';
-import PortalRemotePortalAppRegistryPluginLoader from './PortalRemotePortalAppRegistryPluginLoader';
+import PortalPortalAppRegistryPluginLoader from './PortalPortalAppRegistryPluginLoader';
 
 import type {MashroomPluginLoaderPluginBootstrapFunction} from '@mashroom/mashroom/type-definitions';
 
 const pluginLoaderBootstrap: MashroomPluginLoaderPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
-    return new PortalRemotePortalAppRegistryPluginLoader(context.pluginRegistry, pluginContextHolder.getPluginContext().loggerFactory);
+    return new PortalPortalAppRegistryPluginLoader(context.pluginRegistry, pluginContextHolder.getPluginContext().loggerFactory);
 };
 
 export default pluginLoaderBootstrap;